Make index_offset in turns instead of counts

+++ Firmware/MotorControl/encoder.hpp
@@ -20,7 +20,7 @@ public:
         int32_t phase_offset = 0;        // Offset between encoder count and rotor electrical phase
         float phase_offset_float = 0.0f; // Sub-count phase alignment offset
         int32_t cpr = (2048 * 4);   // Default resolution of CUI-AMT102 encoder,
-        int32_t index_offset = 0;
+        float index_offset = 0.0f;
         uint16_t abs_spi_cs_gpio_pin = 1;
         uint16_t sincos_gpio_pin_sin = 3;
         uint16_t sincos_gpio_pin_cos = 4;
